requirements.txt
# Core LlamaIndex dependencies
llama-index==0.9.30
llama-index-llms-openai==0.1.13
llama-index-embeddings-openai==0.1.7
llama-index-vector-stores-pinecone==0.1.3

# Web scraping and document processing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html2text==2020.1.16

# Document parsing
pypdf2==3.0.1
python-docx==0.8.11
pandas==2.1.4

# Environment management
python-dotenv==1.0.0

# Async processing
aiohttp==3.9.1
asyncio==3.4.3

# Caching
redis==5.0.1
diskcache==5.6.3

# Monitoring and logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# API framework (optional)
fastapi==0.104.1
uvicorn==0.24.0

# Data processing
numpy==1.24.3
scikit-learn==1.3.2

# Progress bars
tqdm==4.66.1
